+++ ReleaseNotes.md
@@ -12,7 +12,7 @@ Released 2023-10-??
 **Updates**
 - MD2/MD3: Dependabot updates
 - MD2/MD3: Migrate all projects to .Net 8 RC 2
-- MD3: Migrate @material/web to 1.0.0-pre.17
+- MD3: Migrate @material/web to 1.0.0
 - MD2: Added new IconFoundry (Material Symbols) and corresponding support and demonstration files
 
 **New components**